bringing the linux-device code upto date with the latest (much-cleaner) watchdog-APIs.

+++ common/instamsg/driver/include/globals.h
@@ -120,5 +120,5 @@ char USER_LOG_FILE_PATH[MAX_BUFFER_SIZE];
 char USER_DEVICE_UUID[MAX_BUFFER_SIZE];
 #endif
 
-unsigned long editableBusinessLogicInterval;
+int editableBusinessLogicInterval;
 #endif
